Saveconfig working and console command 64 bit frequencies

+++ nanovna.h
@@ -582,7 +582,7 @@ typedef struct config {
   int8_t    cor_wfm;
   int8_t    cor_nfm;
   int8_t    ultra;
-  int8_t    dummy;
+  uint32_t    dummy;
 //  uint8_t _reserved[22];
   uint32_t checksum;
 } config_t;
